nav_test: Extract test_site_helper
This will be reused in flatten_url_namespace_test.rb.
